add support to use restify instead of express
This is almost a drop in replacement for express, except it has built
in DTrace support and is a single purpose tool for rest api servers.
With the exception of not supporting the "delete" method (they use "del")
this went off without a hitch. The travis configuration was updated to
test both express and restify