generate endpoints if not provided
The only thing we really need for a resource is a model. This
change allows the user to specify only a model, leaving it up
to epilogue to automatically generate endpoints. This brings
with it a new depedency on node-inflection, however the module
will prove useful again for automatic associations in the near
future.