allow users to disable pagination
A new "pagination" property has been introduced for resources so that
users can conditionally disable pagination for a given resource. A new
test file was introduced for pagination, making it easier to track the
feature in the future. Currently pagination only supports being enabled
or disabled, but the framework is in place to support other options such
as maxCount and resultsPerPage (for example) in the future.