refactor(hooks): reduce work in Controller._control
This dramatically decreases the work done during Controller._control.
Initially milestones were defined for all controllers, including before
and after hooks, regardless of whether they were defined or not. This is
needless work, and thus was removed. The _control method now determines
whether the method was defined in the first place and only runs the work
if it was.